requirements.txt
build==1.0.3
CacheControl==0.13.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
cryptography==42.0.4
distlib==0.3.8
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
idna==3.6
installer==0.7.0
jaraco.classes==3.3.1
jeepney==0.8.0
keyring==24.3.0
more-itertools==10.2.0
msgpack==1.0.7
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pycparser==2.21
pygame-ce==2.4.1
pyproject_hooks==1.0.0
rapidfuzz==3.6.1
requests==2.31.0
requests-toolbelt==1.0.0
ruff==0.2.2
SecretStorage==3.3.3
shellingham==1.5.4
tomlkit==0.12.3
trove-classifiers==2024.2.23
urllib3==2.2.1
virtualenv==20.25.1
